% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.eblupres}
\alias{autoplot.eblupres}
\title{Autoplot EBLUP results}
\usage{
\method{autoplot}{eblupres}(model, variable = "RSE", ...)
}
\arguments{
\item{model}{EBLUP model}

\item{variable}{variable to plot}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Autoplot EBLUP results
}
\examples{
library(saens)
library(ggplot2)

m1 <- eblupfh(y ~ x1 + x2 + x3, data = mys, vardir = "var", cluster = "clust")
autoplot(m1)

}
