% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eblupfh.R
\name{eblupfh}
\alias{eblupfh}
\title{EBLUPs based on a Fay-Herriot Model}
\usage{
eblupfh(
  formula,
  data,
  vardir,
  cluster,
  method = "REML",
  maxiter = 100,
  precision = 1e-04,
  scale = FALSE,
  print_result = TRUE
)
}
\arguments{
\item{formula}{an object of class formula that contains a description of the model to be fitted. The variables included in the formula must be contained in the data.}

\item{data}{a data frame or a data frame extension (e.g. a tibble)}

\item{vardir}{vector or column names from data that contain variance sampling from the direct estimator for each domain.}

\item{cluster}{vector or column name from data that contain cluster information.}

\item{method}{Fitting method can be chosen between 'ML' and 'REML'}

\item{maxiter}{maximum number of iterations allowed in the Fisher-scoring algorithm. Default is 100 iterations.}

\item{precision}{convergence tolerance limit for the Fisher-scoring algorithm. Default value is 0.0001.}

\item{scale}{scaling auxiliary variable or not. Default value is FALSE}

\item{print_result}{print coefficient or not. Default value is TRUE}
}
\value{
List contains data.frame result,
}
\description{
This function gives the EBLUP (or EB predictor under normality) based on a Fay-Herriot model.
}
\examples{
model1 <- eblupfh(y ~ x1 + x2 + x3, data = na.omit(mys), vardir = 'var')
model1 <- eblupfh(y ~ x1 + x2 + x3, data = na.omit(mys), vardir = ~ var)
model2 <- eblupfh(y ~ x1 + x2 + x3, data = mys, vardir = 'var', cluster = 'clust')

}
\references{
- Rao, J. N., & Molina, I. (2015). Small area estimation. John Wiley & Sons.
- Anisa, R., Kurnia, A., & Indahwati, I. (2013). Cluster information of non-sampled area in small area estimation. E-Prosiding Internasional| Departemen Statistika FMIPA Universitas Padjadjaran, 1(1), 69-76.
}
