% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mys}
\alias{mys}
\title{mys: mean years of schooling people with disabilities in Papua Island}
\format{

}
\source{
\url{https://bps.go.id}
}
\usage{
mys
}
\description{
A dataset containing the mean years of schooling people with disabilities in Papua Island in 2021
}
\keyword{datasets}
