% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mys}
\alias{mys}
\title{mys: prediksi harga xxx.}
\format{
A data frame with 57.971 rows and 36 variables:
\describe{
  \item{area}{xxx cost}
  \item{y}{variabel x1}
  \item{var}{variabel x1}
  \item{rse}{variabel x2}
  \item{x1}{date}
  \item{x2}{baris id}
  \item{x3}{baris id}
}
}
\source{
\url{https://apaanini}
}
\usage{
mys
}
\description{
A dataset containing the Cost and other attributes of almost 57.971 observation 36 variabels.
}
\keyword{datasets}
